CRM Proposal Beginner Level
This is a Project Proposal For an Intermediate Project Management Course. This task required that I use the template provided by the Course instructor so that's why the university's watermark is on every page. 
Introduction: 
The proposed solution is a new customer relationship management (CRM) system for the American Video Game Company. There are several disconnected manual and automated process which the new system will be able to manage in a unified way. The proposed system is a custom development and will be able to undergo changes and enhancements that may be required in the future, and it will be a scalable solution. Also, the proposed system will have the ability to integrate with other systems for the improvement of the efficiency of data sharing, reporting, and business process flows.