Proposal for Inventory Management Application
This proposal document was one of the projects required to complete my capstone project for graduation. I previously created an IMS as a university project for a different course, but the story line is make-believe. 

Problem Summary:

Safely Home Security started as a small business selling only one kind of door lock. They have since grown into a company of twenty employees, selling a large variety of customized home security systems. They have a website on which customers fill out a form for information, but they must speak to a sales member to configure a customized security solution, and the sale is made over the phone. The company has a high-ranking trademark reputation for their personalized customer phone support. They are proud to have this interaction versus having the customers go through the difficult process of putting together a security solution and processing an order online.  
However, as sales are increasing, mistakes in the inventory count are occurring frequently, as employees make frequent data entry mistakes because of the rush in sales calls. As a result, Safely Home Security’s employees have been accidently selling out of stock items and having to apologize to customers for these mistakes. They are currently using a manual system of adjusting shared spreadsheets to keep track of inventory. The company has a database, that they update at the end of every day with the results from these spreadsheets. Safely Home Security has recently been receiving complaints and bad reviews, which could lead to the loss of customers.